chore: fix deps
