+++ packages/gestures/__tests__/android/jest.config.js
@@ -1,6 +1,3 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
-const path = require("path");
 
-module.exports = jestConfigurator("android", __dirname, {
-  setupTestFrameworkScriptFile: path.join(__dirname, "./serializers")
-});
+module.exports = jestConfigurator("android", __dirname);
