--- packages/gestures/__tests__/ios/serializers.js
@@ -1,6 +1,0 @@
-const {
-  addSerializers,
-  enzymeRenderedSerializer
-} = require("@times-components/jest-serializer");
-
-addSerializers(expect, enzymeRenderedSerializer());
