+++ packages/gestures/package.json
@@ -41,6 +41,7 @@
     "@times-components/jest-configurator": "2.1.3",
     "@times-components/jest-serializer": "3.0.1",
     "@times-components/storybook": "3.0.1",
+    "@times-components/test-utils": "1.0.0",
     "@times-components/utils": "3.0.1",
     "@times-components/webpack-configurator": "2.0.3",
     "babel-core": "6.26.0",
